pyproject.toml
[project]
name = "solanasearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.59.4",
    "python-dotenv>=1.0.1",
    "toml>=0.10.2",
    "tree-sitter>=0.23.2",
    "tree-sitter-rust>=0.23.2",
    "tree-sitter-languages>=1.10.2",
    "gitpython>=3.1.44",
]
